+++ src/main/java/com/sprint/mission/discodeit/entity/status/AccountStatus.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.entity;
+package com.sprint.mission.discodeit.entity.status;
 
 import java.io.Serializable;
